various fixes after testing previous commits

+++ engine/vision_experiment/gui.py
@@ -8,6 +8,7 @@ import os
 import os.path
 import webbrowser
 import copy
+import traceback
 
 import PyQt4.Qt as Qt
 import PyQt4.QtGui as QtGui
@@ -310,6 +311,7 @@ class RoiWidget(QtGui.QWidget):
         self.show_selected_soma_rois_checkbox = QtGui.QCheckBox(self)
         self.show_selected_roi_centers_label = QtGui.QLabel('Show selected roi centers',  self)
         self.show_selected_roi_centers_checkbox = QtGui.QCheckBox(self)
+        self.show_selected_roi_centers_checkbox.setCheckState(2)
         self.xz_line_length_label = QtGui.QLabel('XZ line length',  self)
         self.xz_line_length_combobox = QtGui.QComboBox(self)
         self.xz_line_length_combobox.setEditable(True)
@@ -324,8 +326,8 @@ class RoiWidget(QtGui.QWidget):
         self.layout = QtGui.QGridLayout()
         self.layout.addWidget(self.roi_info_image_display, 0, 0, 1, 13)
         self.layout.addWidget(self.select_cell_label, 1, 0)
-        self.layout.addWidget(self.select_cell_combobox, 1, 1, 1, 2)
-        self.layout.addWidget(self.cell_id_display_label, 1, 3, 1, 2)
+        self.layout.addWidget(self.select_cell_combobox, 1, 1, 1, 3)
+        self.layout.addWidget(self.cell_id_display_label, 1, 4, 1, 2)
         self.layout.addWidget(self.previous_button, 1, 5)
         self.layout.addWidget(self.accept_cell_button, 1, 6)
         self.layout.addWidget(self.ignore_cell_button, 1, 7)
@@ -656,8 +658,8 @@ class Poller(QtCore.QThread):
                             tag = 'jobhandler'
                         else:
                             tag = parameter
-                        self.backup_mouse_file(tag = tag)
-                        self.queues['analysis']['out'].put('SOCmouse_file_copiedEOCEOP')
+                        if self.backup_mouse_file(tag = tag):
+                            self.queues['analysis']['out'].put('SOCmouse_file_copiedEOCfilename={0}EOP'.format(os.path.split(self.mouse_file)[1].replace('.hdf5', '_jobhandler.hdf5')))
                     else:
                         self.printc(k.upper() + ' '  + message)
 
@@ -683,7 +685,23 @@ class Poller(QtCore.QThread):
             return
         #read cell info from measurement file
         h_measurement = hdf5io.Hdf5io(measurement_file_path)
+        scan_mode = h_measurement.findvar('call_parameters')['scan_mode']
         scan_regions[region_name]['process_status'][id]['find_cells_ready'] = True
+        print scan_mode
+        if scan_mode == 'xz':
+            soma_rois = h_measurement.findvar('soma_rois')
+            if soma_rois is not None:
+                number_of_new_cells = len(soma_rois)
+            else:
+                number_of_new_cells = 0
+            h_measurement.close()
+            #Save changes
+            h.scan_regions = scan_regions
+            h.save(['scan_regions'], overwrite=True)
+            h.close()
+            self.printc('{1} cells found in {0} but not added because it is an xz scan'.format(id, number_of_new_cells))
+            return
+        print 'adding cells'
         h.load('images')
         if not hasattr(h,  'images'):
             h.images = {}
@@ -716,7 +734,7 @@ class Poller(QtCore.QThread):
         else:
             number_of_new_cells = len(soma_rois)
         for i in range(number_of_new_cells):
-            cell_id = ('{0}_{1:2}_{2}_{3}'.format(depth, i, stimulus, id)).replace(' ', '0')
+            cell_id = ('{0}_{1}_{2:2}_{3}'.format(depth, id,  i, stimulus)).replace(' ', '0')
             h.cells[region_name][cell_id] = {}
             h.cells[region_name][cell_id]['depth'] = depth
             h.cells[region_name][cell_id]['id'] = id
@@ -728,7 +746,6 @@ class Poller(QtCore.QThread):
             h.cells[region_name][cell_id]['stimulus'] = stimulus
             h.cells[region_name][cell_id]['roi_curve'] = roi_curves[i]
             h.roi_curves[region_name][cell_id] = roi_curve_images[i]
-            
         h_measurement.close()
         #Save changes
         h.scan_regions = scan_regions
@@ -782,7 +799,7 @@ class Poller(QtCore.QThread):
         measurement_file_path = file.get_measurement_file_path_from_id(id, self.config)
         if measurement_file_path is None or not os.path.exists(measurement_file_path):
             self.printc('Measurement file not found: {0}, {1}' .format(measurement_file_path,  id))
-            return 4*[None]
+            return 5*[None]
         fromfile = hdf5io.read_item(measurement_file_path, ['call_parameters', 'position'])
         call_parameters = fromfile[0]
         depth = fromfile[1]['z'][0]
@@ -790,13 +807,13 @@ class Poller(QtCore.QThread):
         mouse_file = os.path.join(self.config.EXPERIMENT_DATA_PATH, call_parameters['mouse_file'])
         if not os.path.exists(mouse_file):
             self.printc('Mouse file ({0}) assigned to measurement ({1}) is missing' .format(mouse_file,  id))
-            return 4*[None]
+            return 5*[None]
         h = hdf5io.Hdf5io(mouse_file)
         scan_regions = h.findvar('scan_regions')
         if scan_regions[call_parameters['region_name']].has_key(id):
             self.printc('ID already exists: {0}'.format(id))
             h.close()
-            return 4*[None]
+            return 5*[None]
         return scan_regions, call_parameters['region_name'], h, measurement_file_path, depth
  
     def rebuild_cell_database(self):
@@ -824,6 +841,7 @@ class Poller(QtCore.QThread):
         self.cells = {}
         
     def remove_measurement_file_from_database(self, id_to_remove = None, process_status_update = False):
+        self.printc('Removing measurement id...')
         if id_to_remove is None:
             id_to_remove = self.parent.get_current_file_id()
         region_name = self.parent.get_current_region_name()
@@ -856,6 +874,7 @@ class Poller(QtCore.QThread):
             return h
         
     def set_measurement_file_process_state(self):
+        self.printc('Setting state of measurement id...')
         selected_id = self.parent.get_current_file_id()
         target_state = str(self.parent.main_widget.measurement_datafile_status_groupbox.set_to_state_combobox.currentText())
         region_name = self.parent.get_current_region_name()
@@ -880,7 +899,9 @@ class Poller(QtCore.QThread):
         self.backup_mouse_file()
         self.parent.update_jobhandler_process_status(scan_regions)
         self.parent.update_file_id_combobox(scan_regions)
+        self.queues['analysis']['out'].put('SOCclear_joblistEOCEOP')
         self.printc('Measurement file status is updated')
+        
 
     def next_cell(self):
         current_index = self.parent.roi_widget.select_cell_combobox.currentIndex()
@@ -926,6 +947,7 @@ class Poller(QtCore.QThread):
             
     ############## Analysis ########################
     def set_mouse_file(self):
+        return
         self.printc('Mouse file sent to jobhandler')
         self.queues['analysis']['out'].put('SOCset_mouse_fileEOCfilename={0}EOP'.format(os.path.split(self.mouse_file)[1].replace('.hdf5', '_jobhandler.hdf5')))
                                            
@@ -1332,7 +1354,7 @@ class Poller(QtCore.QThread):
         hdf5_handler.close()
         self.backup_mouse_file()
         self.parent.update_region_names_combobox(region_name)
-        self.update_scan_regions()
+        self.update_scan_regions()#This is probably redundant
         self.printc('{0} scan region saved'.format(region_name))
 
     def remove_scan_region(self):
@@ -1770,13 +1792,23 @@ class Poller(QtCore.QThread):
             
     ############# Helpers #############
     def backup_mouse_file(self, mouse_file = None, tag = None):
+        result = False
         if mouse_file is None:
             mouse_file = self.mouse_file
         if tag is None:
             tag = 'copy'
         copy_path = mouse_file.replace('.hdf5', '_' +tag+'.hdf5')
-        shutil.copyfile(mouse_file, copy_path)
+        try:
+            if os.path.exists(mouse_file) and os.path.isfile(mouse_file):
+                shutil.copyfile(mouse_file, copy_path)
+                print 'Trying to open copied hdf5 file'
+                h = hdf5io.Hdf5io(copy_path)#Does not help either
+                h.close()
+                result = True
+        except:
+            self.printc(traceback.format_exc())
         time.sleep(0.2)#Wait to make sure that file is completely copied
+        return result
         
     def create_parameterfile_from_region_info(self, parameter_file_path, scan_type):
         selected_region = self.parent.get_current_region_name()
@@ -1862,17 +1894,6 @@ class Poller(QtCore.QThread):
             return mouse_file_copy
         else:
             return self.mouse_file
-        
-    
-
-# Test cases:
-# 1. move stage - set stage origin - including read stage
-# 2. set / read objective
-# 3. acquire z stack
-# 4. add region with vertical scan test
-# 5. move to region, register and realign
-# 6. vertical realign
-
 
 if __name__ == '__main__':
     pass
